backend/orchestrator/pyproject.toml
[tool.poetry]
name = "orchestrator"
version = "0.1.0"
description = "Orchestrator service for Voice AI Platform"
authors = ["Voice AI Team"]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.104.1"
uvicorn = {extras = ["standard"], version = "^0.24.0"}
grpcio = "^1.58.0"
grpcio-tools = "^1.58.0"
opentelemetry-api = "^1.21.0"
opentelemetry-sdk = "^1.21.0"
opentelemetry-exporter-otlp-proto-grpc = "^1.21.0"
pydantic = "^2.5.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-asyncio = "^0.21.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
